feat: Add response ID to PlaceMapper
The code changes in `PlaceMapper.java` add the `id` field to the `PlaceResponse` object. This change ensures that the response object includes the ID of the place entity.

Based on the recent user commits and repository commits, it seems that the team is working on adding new features to the API related to place descriptions and world cities.